fix: 🐛 break circular dependency causing TDZ error in production builds
Extract types, schemas, and default values from step files into
separate .schema.ts files that don't depend on CreateProjectForm.tsx.

This fixes: ReferenceError: Cannot access 'contributorsStepDefaultValues'
before initialization — caused by Webpack production module ordering
triggering a Temporal Dead Zone between CreateProjectForm.tsx and its
step files, which mutually imported each other.

13 new .schema.ts files created, each containing only the type,
validation schema, and default value exports for one step.
Step .tsx files re-export for backward compatibility.
