+++ components/partials/create/project/steps/ContributorsStep.schema.ts
@@ -1,0 +1,5 @@
+import * as yup from "yup";
+
+export type ContributorsStepValues = Array<string>;
+export const contributorsStepSchema = () => yup.array().of(yup.string().required());
+export const contributorsStepDefaultValues: ContributorsStepValues = [];
