+++ components/partials/create/project/steps/DeclarationsStep.tsx
@@ -17,47 +17,14 @@ import { useFormContext } from "react-hook-form";
 import * as yup from "yup";
 import { CreateProjectValues } from "../CreateProjectForm";
 
-//
+import {
+  type DeclarationsStepValues,
+  declarationsStepSchema,
+  declarationsStepDefaultValues,
+} from "./DeclarationsStep.schema";
+export { type DeclarationsStepValues, declarationsStepSchema, declarationsStepDefaultValues };
 
 const YES_NO = ["yes", "no"] as const;
-
-export interface DeclarationsStepValues {
-  repairable: string;
-  recyclable: string;
-  certifications: Array<ILink>;
-}
-
-yup.setLocale({
-  mixed: {
-    required: "Required",
-  },
-  string: {
-    url: "Invalid URL",
-  },
-});
-
-export const declarationsStepSchema = yup.object({
-  repairable: yup
-    .string()
-    .required()
-    .oneOf([...YES_NO]),
-  recyclable: yup
-    .string()
-    .required()
-    .oneOf([...YES_NO]),
-  certifications: yup.array().of(
-    yup.object().shape({
-      url: yup.string().url().required(),
-      label: yup.string().required(),
-    })
-  ),
-});
-
-export const declarationsStepDefaultValues: DeclarationsStepValues = {
-  repairable: "",
-  recyclable: "",
-  certifications: [],
-};
 
 //
