+++ components/partials/create/project/steps/ImagesStep.schema.ts
@@ -1,0 +1,5 @@
+import * as yup from "yup";
+
+export type ImagesStepValues = Array<File>;
+export const imagesStepSchema = () => yup.array().min(1).required();
+export const imagesStepDefaultValues: ImagesStepValues = [];
