+++ components/partials/create/project/steps/LinkDesignStep.schema.ts
@@ -1,0 +1,5 @@
+import * as yup from "yup";
+
+export type LinkDesignStepValues = string;
+export const linkDesignStepSchema = () => yup.string();
+export const linkDesignStepDefaultValues: LinkDesignStepValues = "";
