+++ components/partials/create/project/steps/MainStep.tsx
@@ -7,41 +7,10 @@ import getIdFromFormName from "lib/getIdFromFormName";
 import { isRequired } from "lib/isFieldRequired";
 import { useTranslation } from "next-i18next";
 import { Controller, useFormContext } from "react-hook-form";
-import * as yup from "yup";
 import { CreateProjectValues, useProjectType } from "../CreateProjectForm";
 
-//
-
-export interface MainStepValues {
-  title: string;
-  description: string;
-  link: string;
-  tags: Array<string>;
-}
-
-export const mainStepDefaultValues: MainStepValues = {
-  title: "",
-  description: "",
-  link: "",
-  tags: [],
-};
-
-export const mainStepSchema = () =>
-  yup.object({
-    title: yup.string().required(),
-    link: yup.string().url().required(),
-    tags: yup.array().of(yup.string()),
-    description: yup
-      .string()
-      .test(
-        "size-check",
-        "Description length must be less than 6000 characters. If it's longer, please use the 'project data' field to reference it.",
-        value => {
-          if (value) return new Blob([value]).size < 6000;
-          else return true;
-        }
-      ),
-  });
+import { type MainStepValues, mainStepDefaultValues, mainStepSchema } from "./MainStep.schema";
+export { type MainStepValues, mainStepDefaultValues, mainStepSchema };
 
 //
