+++ components/partials/create/project/steps/MaterialsStep.tsx
@@ -5,34 +5,9 @@ import SearchMaterials from "components/SearchMaterials";
 import { useTranslation } from "next-i18next";
 import { useState } from "react";
 import { useFormContext } from "react-hook-form";
-import * as yup from "yup";
 import { CreateProjectValues } from "../CreateProjectForm";
 
-export interface MaterialsStepValues {
-  materials: string[]; // Array of material resource IDs
-  materialDetails: Array<{ id: string; name: string }>; // Used for save-time material-* tags
-}
-
-export const materialsStepDefaultValues: MaterialsStepValues = {
-  materials: [],
-  materialDetails: [],
-};
-
-export const materialsStepSchema = () =>
-  yup.object().shape({
-    materials: yup.array().of(yup.string().required()).default([]),
-    materialDetails: yup
-      .array()
-      .of(
-        yup
-          .object({
-            id: yup.string().required(),
-            name: yup.string().required(),
-          })
-          .required()
-      )
-      .default([]),
-  });
+export { type MaterialsStepValues, materialsStepDefaultValues, materialsStepSchema } from "./MaterialsStep.schema";
 
 export default function MaterialsStep() {
   const { t } = useTranslation("createProjectProps");
