+++ components/partials/create/project/steps/ModelFilesStep.schema.ts
@@ -1,0 +1,5 @@
+import * as yup from "yup";
+
+export type ModelFilesStepValues = Array<File>;
+export const modelFilesStepSchema = () => yup.array().default([]);
+export const modelFilesStepDefaultValues: ModelFilesStepValues = [];
