+++ components/partials/create/project/steps/RelationsStep.schema.ts
@@ -1,0 +1,5 @@
+import * as yup from "yup";
+
+export type RelationsStepValues = Array<string>;
+export const relationsStepSchema = () => yup.array().of(yup.string().required());
+export const relationsStepDefaultValues: RelationsStepValues = [];
