+++ components/partials/create/project/steps/ServiceFiltersStep.schema.ts
@@ -1,0 +1,17 @@
+import * as yup from "yup";
+
+export interface ServiceFiltersStepValues {
+  serviceType: string[];
+  availability: string[];
+}
+
+export const serviceFiltersStepDefaultValues: ServiceFiltersStepValues = {
+  serviceType: [],
+  availability: [],
+};
+
+export const serviceFiltersStepSchema = () =>
+  yup.object().shape({
+    serviceType: yup.array().of(yup.string().required()).default([]),
+    availability: yup.array().of(yup.string().required()).default([]),
+  });
